+++ yak_error_handler/lib/src/classes/error_handler/all.dart
@@ -1,2 +1,1 @@
-export 'delegate.dart';
 export 'implementation.dart';