+++ yak_error_handler/lib/src/provider/error_handler.dart
@@ -1,4 +1,4 @@
 import 'package:riverpod/all.dart';
 import '../classes/error_handler/implementation.dart';
 
-final errorHandler = Provider<ErrorHandler>((ref) => ErrorHandler());
+final errorHandler = Provider<YakErrorHandler>((ref) => YakErrorHandler());