+++ yak_error_handler/test/provider/error_handler_test.dart
@@ -13,7 +13,7 @@ void main() {
 
       expect(
         _handler,
-        isA<ErrorHandler>(),
+        isA<YakErrorHandler>(),
         reason: '`read(errorHandler)` should output YakErrorHandler',
       );
     });