+++ yak_test/lib/src/shared/mocks/catch/mock.dart
@@ -1,0 +1,4 @@
+import 'package:mockito/mockito.dart';
+import 'delegate.dart';
+
+class MockCatch extends Mock implements CatchDelegate {}