+++ yak_test/lib/src/yak_runner/test/sync/nullary.dart
@@ -4,7 +4,7 @@ import 'package:meta/meta.dart';
 import 'package:mockito/mockito.dart';
 import 'package:test/test.dart';
 import 'package:yak_runner/yak_runner.dart';
-import 'package:yak_test/src/yak_runner/mocks/all.dart';
+import '../../../shared/all.dart';
 
 class YakRunnerTest<T> implements YakRunnerTestDelegate<T> {
   YakRunnerTest({
@@ -19,7 +19,7 @@ class YakRunnerTest<T> implements YakRunnerTestDelegate<T> {
     assert(result != null, 'result must not be null');
 
     group(description, () {
-      final _errorHandler = MockErrorHandler();
+      final _errorHandler = MockCatch();
       final _delegate = MockDelegate<T>();
       final _runner = YakRunner<T>(_delegate, _errorHandler);
       when(_errorHandler(any, any)).thenAnswer(null);