+++ yak/CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.2.1+1] - 18th January 2021
+
+* bumped all `yak_packages`
+
 ## [0.2.1+0] - 14th January 2021
 
 * added `yak_error_handler`