+++ yak_error_handler/CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.1.0+3] - 18 January 2021
+
+* renaming `ErrorHandler` in `YakErrorHandler`
+
 ## [0.1.0+2] - 17 January 2021
 
 * exporting `ErrorHandler` and `ErrorHandlerDelegate`